fix: bug in task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,4 @@
+#include <stddef.h>		/* defines size_t */
 #include "function_pointers.h"
 /**
  * array_iterator - executes a function given as a parameter
@@ -8,7 +9,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;	/* of type unsigned int */
+	unsigned int i;
 
 	for (i = 0; i < size; i++)
 		action(array[i]);	/* calls func on i'th elem of arr */
